Update README.md
The 42 Philosopher Project is a concurrency and synchronization exercise inspired by the famous "Dining Philosophers Problem" introduced by Edsger Dijkstra. The challenge simulates a group of philosophers who alternately think, eat, and require shared resources (forks) to eat. The project aims to demonstrate how to manage threading, avoid deadlocks, and ensure proper synchronization in a multithreaded environment.

This repository contains my implementation of the 42 Philosopher project, adhering to the rules and constraints provided by the 42 curriculum.